Fix overflow bug and standardize money types to Long
- Fixed bug: entering 9 digits in price field caused reset to 0

- Changed priceForThisTag from Int to Long for consistency

- All money amounts now use Long (incomeCents, totalAmountCents, amountCents, priceForThisTag)

- Added overflow protection in AppointmentEditScreen and ExpenseEditScreen

- Updated database version to 9

- Updated BackupPayload schemaVersion comment to 9

+++ app/src/main/java/com/clientledger/app/data/entity/AppointmentServiceEntity.kt
@@ -32,6 +32,6 @@ data class AppointmentServiceEntity(
     val id: Long = 0,
     val appointmentId: Long,
     val serviceTagId: Long,
-    val priceForThisTag: Int, // in cents, allows override from default
+    val priceForThisTag: Long, // in cents, allows override from default
     val sortOrder: Int = 0 // for ordering tags in UI
 )
